* changed popNumber to long
git-svn-id: file:///usr/local/src/otland/tfssvnrepo/trunk@264 c161f52b-5830-4e7a-bee0-5dea2c17262e

+++ luascript.cpp
@@ -975,10 +975,10 @@ void LuaScriptInterface::popPosition(lua_State* L, Position& position, uint32_t&
 	lua_pop(L, 1); //table
 }
 
-uint32_t LuaScriptInterface::popNumber(lua_State* L)
+uint64_t LuaScriptInterface::popNumber(lua_State* L)
 {
 	lua_pop(L, 1);
-	return (uint32_t)lua_tonumber(L, 0);
+	return (uint64_t)lua_tonumber(L, 0);
 }
 
 double LuaScriptInterface::popFloatNumber(lua_State* L)
